Refactor API imports: separate logs and auth functionalities into distinct modules
